+++ next.config.ts
@@ -15,7 +15,6 @@ const nextConfig: NextConfig = {
         pathname: '/**',
       },
     ],
-    domains: ['sprint-fe-project.s3.ap-northeast-2.amazonaws.com'],
   },
   webpack(config) {
     config.module.rules.push({
