+++ src/api/type/Category.ts
@@ -1,6 +1,6 @@
 export type CategoryResponse = {
   id: number;
   name: string;
-  createdAt: string;
-  updatedAt: string;
+  createdAt?: string;
+  updatedAt?: string;
 };
