+++ src/app/page.tsx
@@ -1,5 +1,11 @@
+import Products from '@/feature/landingpage/Products';
+
 const Page = () => {
-  return <div>공룡</div>;
+  return (
+    <div>
+      <Products />
+    </div>
+  );
 };
 
 export default Page;
