+++ src/feature/landingpage/ProductCard/ProductCard.module.scss
@@ -1,3 +1,74 @@
+.proCardContain {
+  width: 100%;
+  max-width: 300px;
+  background-color: $black-350;
+  border-radius: $radius-common;
+  border: $border-black;
+  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+  overflow: hidden;
+  transition:
+    transform 0.2s ease,
+    box-shadow 0.2s ease;
+
+  &:hover {
+    transform: translateY(-5px);
+    box-shadow: 0 6px 10px rgba(0, 0, 0, 0.15);
+  }
+}
+
+.innerContain {
+  display: flex;
+  flex-direction: column;
+}
+
+.photoContain {
+  width: 100%;
+  height: 150px;
+  overflow: hidden;
+}
+
+.photoContain img {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+}
+
 .innerInfo {
+  padding: 16px;
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+}
+
+.contentTitle {
+  font-size: 20px;
+  font-weight: bold;
   color: $white;
+  margin-bottom: 4px;
+}
+
+.productInfo {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-top: 8px;
+}
+
+.productInfoInner {
+  display: flex;
+  align-items: flex-start;
+  gap: 5px;
+}
+
+.reviewCount,
+.favoriteCount {
+  font-size: 15px;
+  color: $dark-gray;
+}
+
+.raTing {
+  font-size: 15px;
+  font-weight: bold;
+  color: $dark-gray;
+  text-align: right;
 }
