+++ src/feature/landingpage/Products.module.scss
@@ -1,4 +1,5 @@
-.mainTitle,
-.mainTitle2 {
-  color: $white;
+.wrapper {
+  display: flex;
+  flex-direction: column;
+  gap: 30px;
 }
