+++ src/feature/landingpage/ProductsPage/BestProducts.module.scss
@@ -1,0 +1,15 @@
+.topWrapper {
+  display: flex;
+  gap: 10px;
+}
+.mainTitle {
+  color: $white;
+}
+
+.mainTitle2 {
+  color: $gradient;
+}
+.container {
+  display: flex;
+  gap: 20px;
+}
